fix zip generation
